+++ plugin/api/types.go
@@ -2789,7 +2789,61 @@ type TypeReference struct {
 	Name string `json:"name,required"`
 	// Import path for the package defining this type.
 	ImportPath string `json:"importPath,required"`
+	// Annotations defined on this type.
+	//
+	// Note that these are the Thrift annotations listed after the type
+	// declaration in the Thrift file.
+	//
+	// Given,
+	//
+	//   struct User {
+	//     1: required i32 id
+	//     2: required string name
+	//   } (key = "id", validate)
+	//
+	// The annotations will be,
+	//
+	//   {
+	//     "key": "id",
+	//     "validate": "",
+	//   }
+	Annotations map[string]string `json:"annotations,omitempty"`
 }
+
+type _Map_String_String_MapItemList map[string]string
+
+func (m _Map_String_String_MapItemList) ForEach(f func(wire.MapItem) error) error {
+	for k, v := range m {
+		kw, err := wire.NewValueString(k), error(nil)
+		if err != nil {
+			return err
+		}
+
+		vw, err := wire.NewValueString(v), error(nil)
+		if err != nil {
+			return err
+		}
+		err = f(wire.MapItem{Key: kw, Value: vw})
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (m _Map_String_String_MapItemList) Size() int {
+	return len(m)
+}
+
+func (_Map_String_String_MapItemList) KeyType() wire.Type {
+	return wire.TBinary
+}
+
+func (_Map_String_String_MapItemList) ValueType() wire.Type {
+	return wire.TBinary
+}
+
+func (_Map_String_String_MapItemList) Close() {}
 
 // ToWire translates a TypeReference struct into a Thrift-level intermediate
 // representation. This intermediate representation may be serialized
@@ -2808,7 +2862,7 @@ type TypeReference struct {
 //   }
 func (v *TypeReference) ToWire() (wire.Value, error) {
 	var (
-		fields [2]wire.Field
+		fields [3]wire.Field
 		i      int = 0
 		w      wire.Value
 		err    error
@@ -2827,8 +2881,44 @@ func (v *TypeReference) ToWire() (wire.Value, error) {
 	}
 	fields[i] = wire.Field{ID: 2, Value: w}
 	i++
+	if v.Annotations != nil {
+		w, err = wire.NewValueMap(_Map_String_String_MapItemList(v.Annotations)), error(nil)
+		if err != nil {
+			return w, err
+		}
+		fields[i] = wire.Field{ID: 3, Value: w}
+		i++
+	}
 
 	return wire.NewValueStruct(wire.Struct{Fields: fields[:i]}), nil
+}
+
+func _Map_String_String_Read(m wire.MapItemList) (map[string]string, error) {
+	if m.KeyType() != wire.TBinary {
+		return nil, nil
+	}
+
+	if m.ValueType() != wire.TBinary {
+		return nil, nil
+	}
+
+	o := make(map[string]string, m.Size())
+	err := m.ForEach(func(x wire.MapItem) error {
+		k, err := x.Key.GetString(), error(nil)
+		if err != nil {
+			return err
+		}
+
+		v, err := x.Value.GetString(), error(nil)
+		if err != nil {
+			return err
+		}
+
+		o[k] = v
+		return nil
+	})
+	m.Close()
+	return o, err
 }
 
 // FromWire deserializes a TypeReference struct from its Thrift-level
@@ -2872,6 +2962,14 @@ func (v *TypeReference) FromWire(w wire.Value) error {
 				}
 				importPathIsSet = true
 			}
+		case 3:
+			if field.Value.Type() == wire.TMap {
+				v.Annotations, err = _Map_String_String_Read(field.Value.GetMap())
+				if err != nil {
+					return err
+				}
+
+			}
 		}
 	}
 
@@ -2893,14 +2991,35 @@ func (v *TypeReference) String() string {
 		return "<nil>"
 	}
 
-	var fields [2]string
+	var fields [3]string
 	i := 0
 	fields[i] = fmt.Sprintf("Name: %v", v.Name)
 	i++
 	fields[i] = fmt.Sprintf("ImportPath: %v", v.ImportPath)
 	i++
+	if v.Annotations != nil {
+		fields[i] = fmt.Sprintf("Annotations: %v", v.Annotations)
+		i++
+	}
 
 	return fmt.Sprintf("TypeReference{%v}", strings.Join(fields[:i], ", "))
+}
+
+func _Map_String_String_Equals(lhs, rhs map[string]string) bool {
+	if len(lhs) != len(rhs) {
+		return false
+	}
+
+	for lk, lv := range lhs {
+		rv, ok := rhs[lk]
+		if !ok {
+			return false
+		}
+		if !(lv == rv) {
+			return false
+		}
+	}
+	return true
 }
 
 // Equals returns true if all the fields of this TypeReference match the
@@ -2912,6 +3031,9 @@ func (v *TypeReference) Equals(rhs *TypeReference) bool {
 		return false
 	}
 	if !(v.ImportPath == rhs.ImportPath) {
+		return false
+	}
+	if !((v.Annotations == nil && rhs.Annotations == nil) || (v.Annotations != nil && rhs.Annotations != nil && _Map_String_String_Equals(v.Annotations, rhs.Annotations))) {
 		return false
 	}